requirements.txt
torch
phonemizer
inflect
unidecode
scipy
openai
openai-whisper
numpy
pvporcupine
pvcobra
pvrecorder